# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0"
    },
    "paths": {
...
        "/api/visitors": {
            "get": {
                "tags": [
                    "Visitor"
                ],
                "summary": "Список",
                "operationId": "95f1b17a1ca9044cee1bd68f1779154b",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "example@mail.com"
                                                    },
                                                    "first_name": {
                                                        "type": "string",
                                                        "example": "Джон"
                                                    },
                                                    "last_name": {
                                                        "type": "string",
                                                        "example": "Доу"
                                                    },
                                                    "phone": {
                                                        "type": "string",
                                                        "example": "+78005553535"
                                                    },
                                                    "country": {
                                                        "type": "string",
                                                        "example": "Россия"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Токен не предоставлен или срок действия истек или не верный",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Token not provided"
                                                },
                                                "status_code": {
                                                    "type": "integer",
                                                    "example": 401
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Visitor"
                ],
                "summary": "Создание",
                "operationId": "7d8e74a7c2148fdf9bc0a7e1d582304a",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "required": [
                                            "email",
                                            "first_name",
                                            "last_name",
                                            "phone"
                                        ],
                                        "properties": {
                                            "email": {
                                                "type": "string",
                                                "format": "email",
                                                "example": "example@mail.com"
                                            },
                                            "first_name": {
                                                "type": "string",
                                                "example": "Джон"
                                            },
                                            "last_name": {
                                                "type": "string",
                                                "example": "Доу"
                                            },
                                            "phone": {
                                                "type": "string",
                                                "example": "+78005553535"
                                            },
                                            "country": {
                                                "type": "string",
                                                "example": "Россия"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "example@mail.com"
                                                },
                                                "first_name": {
                                                    "type": "string",
                                                    "example": "Джон"
                                                },
                                                "last_name": {
                                                    "type": "string",
                                                    "example": "Доу"
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": "+78005553535"
                                                },
                                                "country": {
                                                    "type": "string",
                                                    "example": "Россия"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Токен не предоставлен или срок действия истек или не верный",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Token not provided"
                                                },
                                                "status_code": {
                                                    "type": "integer",
                                                    "example": 401
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The email field is required."
                                        },
                                        "errors": {
                                            "properties": {
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "example": "The email field is required."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}